+++ react/components/withRuntimeContext.tsx
@@ -1,0 +1,3 @@
+import { withRuntimeContext } from './RenderContext'
+
+export default withRuntimeContext